Create database models Type hint MysqlDb class Create ApiController for RESTFul service

+++ app/libraries/Helpers.php
@@ -9,34 +9,70 @@ use function React\Promise\resolve;
 
 abstract class Helpers
 {
+    /**
+     * Logs an error message to the errors log file
+     * @param string $message
+     * @return void
+     */
     public static function log_error(string $message): void
     {
         $log = new Logger('error');
-        $log->pushHandler(new StreamHandler(dirname(__FILE__, 3) . '/logs/PHP_errors.log', Logger::ERROR));
+        $log->pushHandler(new StreamHandler(ERROR_LOG_FILE, Logger::ERROR));
         $log->error($message);
     }
 
+    /**
+     * Logs an info message to the info log file
+     * @param string $message
+     * @return void
+     *
+     */
     public static function log_info(string $message): void
     {
         $log = new Logger('info');
-        $log->pushHandler(new StreamHandler(dirname(__FILE__, 3) . '/logs/PHP_info.log', Logger::INFO));
+        $log->pushHandler(new StreamHandler(INFO_LOG_FILE, Logger::INFO));
         $log->info($message);
     }
 
+    /**
+     * Encrypts the given OTP
+     * @param string $otp
+     * @return string
+     */
     public static function encrypt_otp(string $otp): string
     {
         return password_hash($otp, PASSWORD_DEFAULT);
     }
 
+    /**
+     * Verifies the given OTP
+     * @throws Exception
+     */
+    public static function verify_otp(string $otp): bool
+    {
+        // Has the OTP expired?
+        if (time() > self::fetch_session_data(SessionKeys::OTP_EXPIRY)) {
+            return false;
+        }
+        return password_verify($otp, self::fetch_session_data(SessionKeys::OTP));
+    }
+
+    /**
+     * Adds the given value to the session
+     * @param string $key
+     * @param mixed $value
+     * @return void
+     */
     public static function add_to_session(string $key, mixed $value): void
     {
         $_SESSION[$key] = $value;
     }
 
     /**
+     * Fetches the value of the given key from the session
      * @throws Exception
      */
-    public static function fetch_from_session(string $key): mixed
+    public static function fetch_session_data(string $key): mixed
     {
         if (!isset($_SESSION[$key])) {
             throw new Exception(ExceptionType::INVALID_SESSION_KEY);
@@ -44,79 +80,59 @@ abstract class Helpers
         return $_SESSION[$key];
     }
 
-    public static function verify_otp(string $otp): bool
-    {
-        // Has the OTP expired?
-        if (time() > self::fetch_from_session(SessionKeys::OTP_EXPIRY)) {
-            return false;
-        }
-        return password_verify($otp, self::fetch_from_session(SessionKeys::OTP));
-    }
-
     public static function is_logged_in(): bool
     {
-        return isset($_SESSION['logged_in_user']);
+        // Logged_in_user is set and not null
+        try {
+            $logged_in_user = self::get_logged_in_user();
+            return !is_null($logged_in_user);
+        } catch (Exception $e) {
+            self::log_error($e->getMessage());
+            return false;
+        }
     }
 
-    // Get role from session and check if it is admin
+    /**
+     * Checks if the logged-in user is an admin
+     */
     public static function is_admin(): bool
     {
-        return self::getLoggedInUserRole() == 'admin';
+        try {
+            $logged_in_user = self::get_logged_in_user();
+            if (is_null($logged_in_user)) {
+                return false;
+            }
+            return $logged_in_user->isAdmin();
+        } catch (Exception $e) {
+            self::log_error($e->getMessage());
+            return false;
+        }
     }
 
-    // Get role from session and check if it is student
-    public static function is_student(): bool
-    {
-        return self::getLoggedInUserRole() == 'student';
-    }
+    /*    #[NoReturn] public static function redirect(string $relativeURL = ''): void
+        {
+            $url = self::parseURL($relativeURL);
+            header('Location: ' . $url);
+            exit;
+        }*/
 
-    #[NoReturn] public static function redirect(string $relativeURL = ''): void
+    #[NoReturn] public static function redirect_to($url): void
     {
-        $url = self::parseURL($relativeURL);
         header('Location: ' . $url);
         exit;
     }
 
-    #[NoReturn] public static function redirectTo($url): void
-    {
-        header('Location: ' . $url);
-        exit;
-    }
-
-    public static function redirect_with_params(string $controller, string $method = '', string $params = ''): void
-    {
-        $url = self::parseURL($controller, $method) . '?' . $params;
-        header('location: ' . $url);
-    }
-
-    public static function get_request_method(): string
-    {
-        return $_SERVER['REQUEST_METHOD'];
-    }
-
-    public static function is_post_method(): bool
-    {
-        return $_SERVER['REQUEST_METHOD'] == 'POST';
-    }
-
-    public static function is_get_method(): bool
-    {
-        return $_SERVER['REQUEST_METHOD'] == 'GET';
-    }
+    /*    public static function redirect_with_params(string $controller, string $method = '', string $params = ''): void
+        {
+            $url = self::parseURL($controller, $method) . '?' . $params;
+            header('location: ' . $url);
+        }*/
 
     public static function concat_string(string $separator, ...$args): string
     {
         $args = func_get_args();
         $parts = array_slice($args, 1);
         return implode($separator, $parts);
-    }
-
-    public static function parseURL(string ...$url_parts): string
-    {
-        $url_parts = func_get_args();
-        $processed_args = array_map(fn($arg) => trim($arg, '/'), $url_parts);
-        $url_root = rtrim(URL_ROOT, '/');
-        return $url_root . '/' . implode('/', $processed_args);
     }
 
     /**
@@ -126,7 +142,7 @@ abstract class Helpers
     {
         // Check if $field is set in $_POST
         if (!isset($_POST[$field])) {
-            throw new Exception(ExceptionType::INVALID_POST_DATA_FIELD . ' ' . $field );
+            throw new Exception(ExceptionType::INVALID_POST_DATA_FIELD . ' ' . $field);
         }
         // Validate post data using self::validate_post_data() function
         $validated = self::validate_post_data($field);
@@ -189,7 +205,7 @@ abstract class Helpers
     /**
      * Flash message helper.
      *
-     * EXAMPLE: flash('register_success', 'You are now registered');
+     * EXAMPLE: flash('register_success' 'You are now registered');
      *
      * DISPLAY IN VIEW: echo flash('register_success');
      * @param string $title
@@ -217,17 +233,8 @@ html;
         }
     }
 
-
-    public static function isLoggedIn(): bool
-    {
-        if (isset($_SESSION['logged_in_user'])) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
     /**
+     * Generates a random string
      * @throws Exception
      */
     public static function generate_otp(): string
@@ -236,13 +243,14 @@ html;
     }
 
     /**
+     * Sends an email using the given PHPMailer object or creates a new one if none is provided
      * @throws Exception
      */
     public static function send_email(EmailModel $email, ?PHPMailer $mailer = null): bool
     {
 
         // Set up the PHPMailer object
-        $mail = $mailer ?? self::configureMailer();
+        $mail = $mailer ?? self::configure_mailer();
         // Turn on SMTP debugging
         $mail->SMTPDebug = SMTP::DEBUG_SERVER;
         try {
@@ -265,9 +273,10 @@ html;
     }
 
     /**
+     * Configures the PHPMailer object for sending emails
      * @throws \PHPMailer\PHPMailer\Exception
      */
-    public static function configureMailer(): PHPMailer
+    public static function configure_mailer(): PHPMailer
     {
         $mailer = new PHPMailer(true);
         $mailer->SMTPDebug = SMTP::DEBUG_CLIENT;
@@ -290,6 +299,9 @@ html;
     }
 
 
+    /**
+     *  Destroys the current session
+     */
     public static function destroy_session(): bool
     {
         // Clear session data and destroy session
@@ -299,11 +311,12 @@ html;
 
 
     /**
+     * Sends an OTP to the given email address
      * @throws Exception
      */
-    public static function send_otp_to_email($email, $otp): void
+    public static function send_otp_to_email(string $email, string $otp): void
     {
-        // Create child process using pcntl_fork to send OTP to email asynchronously
+        // Create child process using pcntl_fork to send OTP to email asynchronously and i
         $pid = pcntl_fork();
         if ($pid === -1) {
             // Error: failed to fork
@@ -314,7 +327,7 @@ html;
             // Create an event loop
             $loop = React\EventLoop\Loop::get();
             // create email model
-            $email_model = EmailModel::factory([
+            $email_model = new EmailModel([
                 EmailModelSchema::RECIPIENT_ADDRESS => $email,
                 EmailModelSchema::SUBJECT => 'Login Code',
                 EmailModelSchema::BODY => "Your login code is: $otp"
@@ -341,66 +354,15 @@ html;
             $loop->run();
 
             // Properly end the child process using posix_kill
-            posix_kill(posix_getpid(), SIGTERM);
-        } /*else {
-            // Parent process: continue executing the main application
-            // Add to encrypted otp to session
-            $otp_encrypted = Helpers::encryptOTP($otp);
-            Helpers::add_to_session(SessionKeys::OTP_ENCRYPTED, $otp_encrypted);
-        }*/
-        // close current session
-        // session_write_close();
+            //posix_kill(posix_getpid(), SIGTERM);
+            exit;
+        }
     }
-
-    // Function to send json_encoded output. Use JSON_PRETTY_PRINT. Example:
-    #[NoReturn] public static function sendJson(array $data, int $status = 200, bool $pretty = true): void
-    {
-        header('Content-Type: application/json');
-        http_response_code($status);
-        echo json_encode($data, $pretty ? JSON_PRETTY_PRINT : 0);
-        exit;
-    }
-
-    // Function to get the currently logged-in user from session. Returns false if no user is logged in.
-    public static function getLoggedInUser(): UserModel|bool
-    {
-        return $_SESSION['logged_in_user'] ?? false;
-    }
-
-    // Function to get the currently logged-in user's id from session. Returns false if no user is logged in.
-    public static function getLoggedInUserId(): int|bool
-    {
-        return $_SESSION['logged_in_user']->id ?? false;
-    }
-
-    // Function to get the currently logged-in user's role from session. Returns false if no user is logged in.
-    public static function getLoggedInUserRole(): string|bool
-    {
-        return $_SESSION['logged_in_user']->role ?? false;
-    }
-
 
     #[NoReturn] public static function logout(): void
     {
         // Destroy session
         Helpers::destroy_session();
-    }
-
-    /**
-     * @throws Exception
-     */
-    public static function get_from_session(string $key): mixed
-    {
-        // Check if $key is set in $_SESSION
-        if (!isset($_SESSION[$key])) {
-            throw new Exception(ExceptionType::INVALID_SESSION_KEY);
-        }
-        return $_SESSION[$key];
-    }
-
-    public static function is_valid_otp(string $otp): bool
-    {
-        return self::verify_otp($otp);
     }
 
     public static function remove_from_session(string $key): void
@@ -409,17 +371,17 @@ html;
     }
 
     /**
+     * Logs in a user using their email and OTP
      * @throws Exception
      */
     public static function login(string $email, bool $verify_otp): void
     {
-        // Todo: Check if email exists in database
-        /*if (!UserModel::emailExists($email)) {
+        // Check if email exists in database
+        if (!UserModel::emailExists($email)) {
             echo json_encode(array('success' => false, 'message' => 'Email address not found!'));
             return;
-        }*/
-
-        if(!$verify_otp) {
+        }
+        if (!$verify_otp) {
 
             // Add email to session
             Helpers::add_to_session(SessionKeys::EMAIL, $email);
@@ -437,30 +399,30 @@ html;
             Helpers::add_to_session(SessionKeys::OTP_EXPIRY, time() + 5 * 60);
 
             // Send response to client
-            echo json_encode(array('success' => true, 'message' => 'We have sent a login code to your email address.'));
+            echo json_encode(array('success' => true, 'message' => 'A login code should be sent to your email address shortly! <br /> Please check your email and enter the code below to login!'));
         } else {
             // Verify user submitted OTP
             // Get OTP from POST data
             $otp = Helpers::fetch_post_data('otp');
 
             // Check if OTP is valid
-            if (!Helpers::is_valid_otp($otp)) {
+            if (!Helpers::verify_otp($otp)) {
                 echo json_encode(array('success' => false, 'message' => 'Invalid or Expired OTP!'));
                 return;
             }
             // Get email from session
-            $email = Helpers::get_from_session(SessionKeys::EMAIL);
+            $email_from_session = Helpers::fetch_session_data(SessionKeys::EMAIL);
+            // Check if email from session is same as email from POST data
+            if ($email_from_session !== $email) {
+                echo json_encode(array('success' => false, 'message' => 'Invalid email address!'));
+                return;
+            }
 
-            // Todo: Get user details from database
-            // $user = UserModel::getUserByEmail($email);
-
-            // Todo: Add user details to session
-//            Helpers::add_to_session(SessionKeys::USER_ID, $user->id);
-//            Helpers::add_to_session(SessionKeys::USER_NAME, $user->name);
-//            Helpers::add_to_session(SessionKeys::USER_EMAIL, $user->email);
+            // Get user details from database
+            $user = UserModel::getUserByEmail($email);
 
             // Set is logged in to true
-            Helpers::add_to_session(SessionKeys::LOGGED_IN_USER, true);
+            Helpers::add_to_session(SessionKeys::LOGGED_IN_USER, $user);
 
             // Send response to client
             echo json_encode(array('success' => true, 'message' => 'Login successful!'));
@@ -473,24 +435,180 @@ html;
         }
     }
 
-    private static function is_async_supported(): bool
+    /**
+     * Gets the logged-in user from the session
+     */
+    public static function get_logged_in_user(): UserModel|null
     {
-        return extension_loaded('pcntl') && extension_loaded('posix');
+        try {
+            return Helpers::fetch_session_data(SessionKeys::LOGGED_IN_USER);
+        } catch (Exception $e) {
+            Helpers::log_error($e->getMessage());
+            return null;
+        }
     }
 
+    /**
+     *  Returns a json encoded string of the data
+     * @param array $data
+     * @return string|false
+     */
+    public static function json_encode(array $data): string|false
+    {
+        return json_encode($data, JSON_PRETTY_PRINT);
+    }
+
+    #[NoReturn]
+    public static function redirect_to_404(): void
+    {
+        header('Location: ' . URL_ROOT . '/errors/index/404', response_code: 302); // 302 is used to redirect browsers to the 404 page
+        exit;
+    }
+
+    /**
+     * Sends an HTTP response with the given status code and message
+     */
+    public static function http_response_code(int $code, string $message = null, string $contentType = 'application/json'): void
+    {
+        // Get the HTTP status text that corresponds to the code
+        $statusText = self::getHttpStatusText($code);
+
+        // Set the HTTP response code and status text
+        header("HTTP/1.1 $code $statusText", true, $code);
+
+        // Set the content type
+        header("Content-Type: $contentType");
+
+        // If a custom message is provided, send it as the response body
+        if ($message !== null) {
+            echo $message;
+        }
+    }
+
+    /**
+     * Returns the HTTP status text that corresponds to the given code
+     */
+    private static function getHttpStatusText(int $code): string
+    {
+        $statusTexts = [
+            200 => 'OK',
+            400 => 'Bad Request',
+            401 => 'Unauthorized',
+            403 => 'Forbidden',
+            404 => 'Not Found',
+            500 => 'Internal Server Error',
+            // Add other status codes as needed
+        ];
+
+        return $statusTexts[$code] ?? 'Unknown Status';
+    }
+
+    public static function is_valid_date(string $datetime): bool
+    {
+        return strtotime($datetime) !== false;
+    }
+
+    /**
+     * Gets the content type of the request
+     * @return string
+     */
+    public static function get_content_type(): string
+    {
+        return $_SERVER['CONTENT_TYPE'];
+    }
+
+    /**
+     * Get the request body as an associative array if the Content-Type header is application/json
+     * @throws Exception
+     */
+    public static function get_json_data()
+    {
+        // Get the raw POST data
+        $rawData = file_get_contents('php://input');
+
+        // Decode the JSON data into an associative array
+        $jsonData = json_decode($rawData, true);
+
+        // Check if the JSON is valid
+        if (json_last_error() !== JSON_ERROR_NONE) {
+            // Handle the error as needed
+            throw new Exception('Invalid JSON input');
+        }
+
+        return $jsonData;
+    }
+
+    /**
+     * Get the request body as an associative array
+     * @return array
+     */
+    public static function get_post_data(): array
+    {
+        return $_POST;
+    }
+
+    /**
+     * Checks if the http request method is PUT
+     * @return bool
+     */
+    public static function is_put(): bool
+    {
+        return $_SERVER['REQUEST_METHOD'] === 'PUT';
+    }
+
+    /**
+     * Checks if the http request method is PATCH
+     * @return bool
+     */
+    public static function is_patch(): bool
+    {
+        return $_SERVER['REQUEST_METHOD'] === 'PATCH';
+    }
+
+    /**
+     * Checks if the http request method is DELETE
+     * @return bool
+     */
+    public static function is_delete(): bool
+    {
+        return $_SERVER['REQUEST_METHOD'] === 'DELETE';
+    }
+
+    /**
+     * Checks if the http request method is POST
+     * @return bool
+     */
     public static function is_post(): bool
     {
         return $_SERVER['REQUEST_METHOD'] == 'POST';
     }
 
+    /**
+     * Checks if the http request method is GET
+     * @return bool
+     */
     public static function is_get(): bool
     {
         return $_SERVER['REQUEST_METHOD'] == 'GET';
     }
 
+    /**
+     * Checks if the request is an AJAX request
+     * @return bool
+     */
     public static function is_ajax(): bool
     {
         return isset($_SERVER['HTTP_X_REQUESTED_WITH']) && strtolower($_SERVER['HTTP_X_REQUESTED_WITH']) == 'xmlhttprequest';
+    }
+
+    /**
+     * Adds a location header to the response
+     * @param string $string
+     * @return void
+     */
+    public static function add_location_header(string $string): void
+    {
+        header('Location: ' . $string);
     }
